+++ x/fees/module.go
@@ -1,0 +1,171 @@
+package fees
+
+import (
+	"encoding/json"
+	"math/rand"
+
+	"github.com/cosmos/cosmos-sdk/client/context"
+	"github.com/cosmos/cosmos-sdk/codec"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/module"
+	"github.com/cosmos/cosmos-sdk/x/auth"
+	sim "github.com/cosmos/cosmos-sdk/x/simulation"
+	"github.com/desmos-labs/desmos/x/fees/client/cli"
+	"github.com/desmos-labs/desmos/x/fees/client/rest"
+	"github.com/desmos-labs/desmos/x/fees/keeper"
+	"github.com/desmos-labs/desmos/x/fees/simulation"
+	"github.com/desmos-labs/desmos/x/fees/types"
+	"github.com/gorilla/mux"
+	"github.com/spf13/cobra"
+	abci "github.com/tendermint/tendermint/abci/types"
+)
+
+// type check to ensure the interface is properly implemented
+var (
+	_   module.AppModule           = AppModule{}
+	_   module.AppModuleBasic      = AppModuleBasic{}
+	_   module.AppModuleSimulation = AppModule{}
+	cdc                            = codec.New()
+)
+
+// AppModuleBasic defines the basic application module used by the posts module.
+type AppModuleBasic struct{}
+
+// Name returns the posts module's name.
+func (AppModuleBasic) Name() string {
+	return types.ModuleName
+}
+
+// RegisterCodec registers the posts module's types for the given codec.
+func (AppModuleBasic) RegisterCodec(cdc *codec.Codec) {}
+
+// DefaultGenesis returns default genesis state as raw bytes for the auth
+// module.
+func (AppModuleBasic) DefaultGenesis() json.RawMessage {
+	return cdc.MustMarshalJSON(types.DefaultGenesisState())
+}
+
+// ValidateGenesis performs genesis state validation for the posts module.
+func (AppModuleBasic) ValidateGenesis(bz json.RawMessage) error {
+	var data types.GenesisState
+	err := cdc.UnmarshalJSON(bz, &data)
+	if err != nil {
+		return err
+	}
+	// Once json successfully marshalled, passes along to genesis.go
+	return types.ValidateGenesis(data)
+}
+
+// RegisterRESTRoutes registers the REST routes for the posts module.
+func (AppModuleBasic) RegisterRESTRoutes(ctx context.CLIContext, rtr *mux.Router) {
+	rest.RegisterRoutes(ctx, rtr)
+}
+
+// GetTxCmd returns the root tx command for the posts module.
+func (AppModuleBasic) GetQueryCmd(cdc *codec.Codec) *cobra.Command {
+	return cli.GetQueryCmd(cdc)
+}
+
+// GetQueryCmd returns the root query command for the posts module.
+func (AppModuleBasic) GetTxCmd(_ *codec.Codec) *cobra.Command {
+	return nil
+}
+
+//____________________________________________________________________________
+
+// AppModule implements an application module for the posts module.
+type AppModule struct {
+	AppModuleBasic
+	ak     auth.AccountKeeper
+	keeper keeper.Keeper
+}
+
+// NewAppModule creates a new AppModule Object
+func NewAppModule(keeper keeper.Keeper, accountKeeper auth.AccountKeeper) AppModule {
+	return AppModule{
+		AppModuleBasic: AppModuleBasic{},
+		ak:             accountKeeper,
+		keeper:         keeper,
+	}
+}
+
+// Name returns the posts module's name.
+func (AppModule) Name() string {
+	return types.ModuleName
+}
+
+// RegisterInvariants performs a no-op.
+func (am AppModule) RegisterInvariants(_ sdk.InvariantRegistry) {}
+
+// Route returns the message routing key for the posts module.
+func (am AppModule) Route() string {
+	return types.RouterKey
+}
+
+// NewHandler returns an sdk.Handler for the posts module.
+func (am AppModule) NewHandler() sdk.Handler {
+	return nil
+}
+
+// QuerierRoute returns the posts module's querier route name.
+func (am AppModule) QuerierRoute() string {
+	return types.QuerierRoute
+}
+
+// NewQuerierHandler returns the posts module sdk.Querier.
+func (am AppModule) NewQuerierHandler() sdk.Querier {
+	return keeper.NewQuerier(am.keeper)
+}
+
+// InitGenesis performs genesis initialization for the posts module. It returns
+// no validator updates.
+func (am AppModule) InitGenesis(ctx sdk.Context, data json.RawMessage) []abci.ValidatorUpdate {
+	var genesisState types.GenesisState
+	cdc.MustUnmarshalJSON(data, &genesisState)
+	return InitGenesis(ctx, am.keeper, genesisState)
+}
+
+// ExportGenesis returns the exported genesis state as raw bytes for the auth
+// module.
+func (am AppModule) ExportGenesis(ctx sdk.Context) json.RawMessage {
+	gs := ExportGenesis(ctx, am.keeper)
+	return cdc.MustMarshalJSON(gs)
+}
+
+// BeginBlock returns the begin blocker for the posts module.
+func (am AppModule) BeginBlock(_ sdk.Context, _ abci.RequestBeginBlock) {
+}
+
+// EndBlock returns the end blocker for the posts module. It returns no validator
+// updates.
+func (am AppModule) EndBlock(_ sdk.Context, _ abci.RequestEndBlock) []abci.ValidatorUpdate {
+	return []abci.ValidatorUpdate{}
+}
+
+//____________________________________________________________________________
+
+// AppModuleSimulation defines the module simulation functions used by the posts module.
+type AppModuleSimulation struct{}
+
+// GenerateGenesisState creates a randomized GenState of the bank module.
+func (AppModule) GenerateGenesisState(simState *module.SimulationState) {
+	simulation.RandomizedGenState(simState)
+}
+
+// ProposalContents doesn't return any content functions for governance proposals.
+func (AppModule) ProposalContents(_ module.SimulationState) []sim.WeightedProposalContent {
+	return nil
+}
+
+// RandomizedParams creates randomized posts param changes for the simulator.
+func (AppModule) RandomizedParams(r *rand.Rand) []sim.ParamChange {
+	return simulation.ParamChanges(r)
+}
+
+// RegisterStoreDecoder performs a no-op.
+func (AppModule) RegisterStoreDecoder(sdr sdk.StoreDecoderRegistry) {}
+
+// WeightedOperations returns the all the posts module operations with their respective weights.
+func (am AppModule) WeightedOperations(simState module.SimulationState) []sim.WeightedOperation {
+	return nil
+}